requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0

# Security
bandit>=1.7.0
safety>=3.0.0
pip-audit>=2.7.0

# Linting
ruff>=0.1.0

# YDB
ydb>=3.0.0

# S3
boto3>=1.34.0

# HTTP client
httpx>=0.27.0

# Telegram
python-telegram-bot>=21.0
